factored datastore out into a module
