+++ src/datastore/mod.rs
@@ -1,0 +1,9 @@
+pub mod delta;
+pub mod history;
+pub mod branch;
+pub mod datastore;
+
+pub use delta::Delta;
+pub use history::History;
+pub use branch::Branch;
+pub use datastore::Datastore;
